+++ junit/spring-integration-test/src/test/java/com/fitbur/testify/integration/ProviderRealInjectionTest.java
@@ -28,7 +28,7 @@ import org.mockito.internal.util.MockUtil;
  *
  * @author saden
  */
-@RunWith(SpringIntegrationTestRunner.class)
+@RunWith(SpringIntegrationTest.class)
 @Module(SpringIntegrationConfig.class)
 public class ProviderRealInjectionTest {