+++ CHANGELOG.md
@@ -13,8 +13,6 @@ was added, changed, deprecated, removed, fix and security fixes.
 
 ## [Unreleased]
 
-### Changed
-- Changed Mock annotation to Fake
 
 ## [0.0.1] - 2015-11-23
 ### Added
@@ -56,4 +54,13 @@ was added, changed, deprecated, removed, fix and security fixes.
 - Removed TestException class in favor of IllegalStateException
 - Removed HSQLTestException and HSQLContext (not necessary with the addition of NeedProvider)
 - Removed Tests module in favor of testing inside modules themselves
+
+## [0.0.4] - 2015-12-10
+### Changed
+- Changed Mock annotation to Fake
+
+## [0.0.5] - 2015-12-10
+- Changed UnitTestRunner to UnitTest
+- Changed SpringIntegrationTestRunner to SpringIntegrationTest
+- Host of fixes for issue relating to test class injection